remove unneeded arguments check
All advice functions should be used in a predetermined manner
and only for predetermined purposes.
It is therefore reasonable not to check the arguments.

As proof of this, no other advising function performs the same checks.
Remove the non-essential code and focus only on what we want to focus on.